Primary field improvement
Primary field of table is now the first one to be iterated when calculating
cartisian product of fields. That means primary field will only be iterated
once as it is the highest level of loop.

This also gives more consistent records as they will all flow based on items
of primary field.

'.get_other_fields()' method was also added to table and tests updated.